Split compositeRuntime to its own subdirectory
