+++ examples/utils/migration-tools/src/compositeRuntime/index.ts
@@ -1,0 +1,10 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export { IEntryPointPiece } from "./interfaces.js";
+export {
+	CompositeEntryPoint,
+	loadCompositeRuntime,
+} from "./loadCompositeRuntime.js";
