+++ examples/utils/migration-tools/src/migratableModelLoader/migrationToolEntryPointPiece.ts
@@ -6,10 +6,9 @@
 import type { IContainerRuntime } from "@fluidframework/container-runtime-definitions/internal";
 import type { IFluidHandle } from "@fluidframework/core-interfaces";
 
+import type { IEntryPointPiece } from "../compositeRuntime/index.js";
 import type { IMigrationTool } from "../interfaces/index.js";
 import { MigrationToolFactory } from "../migrationTool.js";
-
-import type { IEntryPointPiece } from "./loadCompositeRuntime.js";
 
 const migrationToolEntryPointPieceName = "migrationTool";
